requirements.txt
setuptools
python-multipart
fastapi
uvicorn[standard]
httpx
openai
redis
python-dotenv